# Add wasm plugin benchmark